bench-mark with fortran with fluctuation on
when add a fluctuation, the two code runs exactly the same too. Output Bz and rhoVx vaiable, and compare them using matlab, giving a difference in 10^-15 which is determined by Standard-Output of Computer Language. As to now, we can absolutly trust the c++ code, which will give an exactly same results with fortran code.